pyproject.toml
[project]
name = "task-master"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "django>=5.0.7",
    "psycopg2-binary>=2.9.9",
    "django-model-utils>=4.5.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/task_master"]
